đang sửa cái payos

+++ src/main/java/com/example/demo/dto/response/OrderItemResponse.java
@@ -5,6 +5,6 @@ import java.util.UUID;
 public record OrderItemResponse(
         UUID productId,
         int quantity,
-        Double price
+        Long price
 ) {
 }
